Find references should not open files not already open.
We do this solely to determine the preview of line content,
and it can be very expensive to open many files on a larger
project (which can even reach into the standard library).

Instead we just show "..." for content for any file not already
open.  This isn't optimal, but it's what we can easily do without
paying a large penalty.